Conflicts:
	GalCon-Client/GalCon-Client-android/gen/com/xxx/galcon/R.java

+++ GalCon-Client/GalCon-Client/src/com/xxx/galcon/http/JsonConstructor.java
@@ -47,7 +47,7 @@ public class JsonConstructor {
 		return top;
 	}
 	
-	public static JSONObject addCoins(String playerHandle, Long numCoins, Long usedCoins) throws JSONException{
+	public static JSONObject addCoins(String playerHandle, int numCoins, Long usedCoins) throws JSONException{
 		JSONObject top = new JSONObject();
 		
 		top.put("playerHandle", playerHandle);
